+++ ceno_emul/tests/test_elf.rs
@@ -1,5 +1,8 @@
 use anyhow::Result;
-use ceno_emul::{ByteAddr, CENO_PLATFORM, EmuContext, InsnKind, Platform, StepRecord, VMState};
+use ceno_emul::{
+    CENO_PLATFORM, EmuContext, InsnKind, Platform, StepRecord, VMState,
+    host_utils::read_all_messages,
+};
 
 #[test]
 fn test_ceno_rt_mini() -> Result<()> {
@@ -65,10 +68,10 @@ fn test_ceno_rt_io() -> Result<()> {
 
     let all_messages = read_all_messages(&state);
     for msg in &all_messages {
-        print!("{}", String::from_utf8_lossy(msg));
+        print!("{msg}");
     }
-    assert_eq!(&all_messages[0], "📜📜📜 Hello, World!\n".as_bytes());
-    assert_eq!(&all_messages[1], "🌏🌍🌎\n".as_bytes());
+    assert_eq!(&all_messages[0], "📜📜📜 Hello, World!\n");
+    assert_eq!(&all_messages[1], "🌏🌍🌎\n");
     Ok(())
 }
 
@@ -76,35 +79,4 @@ fn run(state: &mut VMState) -> Result<Vec<StepRecord>> {
     let steps = state.iter_until_halt().collect::<Result<Vec<_>>>()?;
     eprintln!("Emulator ran for {} steps.", steps.len());
     Ok(steps)
-}
-
-const WORD_SIZE: usize = 4;
-const INFO_OUT_ADDR: u32 = 0xC000_0000;
-
-fn read_all_messages(state: &VMState) -> Vec<Vec<u8>> {
-    let mut all_messages = Vec::new();
-    let mut word_offset = 0;
-    loop {
-        let out = read_message(state, word_offset);
-        if out.is_empty() {
-            break;
-        }
-        word_offset += out.len().div_ceil(WORD_SIZE) as u32 + 1;
-        all_messages.push(out);
-    }
-    all_messages
-}
-
-fn read_message(state: &VMState, word_offset: u32) -> Vec<u8> {
-    let out_addr = ByteAddr(INFO_OUT_ADDR).waddr() + word_offset;
-    let byte_len = state.peek_memory(out_addr);
-    let word_len_up = byte_len.div_ceil(4);
-
-    let mut info_out = Vec::with_capacity(WORD_SIZE * word_len_up as usize);
-    for i in 1..1 + word_len_up {
-        let value = state.peek_memory(out_addr + i);
-        info_out.extend_from_slice(&value.to_le_bytes());
-    }
-    info_out.truncate(byte_len as usize);
-    info_out
 }
